% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_shape.R
\name{cut_shape}
\alias{cut_shape}
\title{Cut some shapes!}
\usage{
cut_shape(image, shape, color = "#000000", fuzz = 10)
}
\arguments{
\item{image}{The image from which the shape will be cut}

\item{shape}{The image that contains the shape to cut out}

\item{color}{The color, as a hex code (e.g. #000000), that identifies the shape in the image
This is black by default}

\item{fuzz}{The fuzz factor to identify the points that match the color specified.
Increase to get a smoother outline for the cut shape.}
}
\value{
An external pointer of class 'magick-image'
}
\description{
If you can make or find a shape as an image, you can use it cut out shapes from other images.
}
\examples{
cut_shape(image = "https://cdn.pixabay.com/photo/2019/09/08/19/01/pumpkin-4461665_1280.jpg",
shape = "https://imageog.flaticon.com/icons/png/512/30/30209.png?size=1200x630f&pad=10,10,10,10&ext=png&bg=FFFFFFFF")
}
